# services/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLPRouter services API",
    "version": "0.1.0",
    "description": "Route status, registry views, quotes and the public forward/flush trigger queue of CLPRouter. Optional and re-checkable: every answer carries the block numbers (`inputs`) and registry versions it was built from. Read-only, unauthenticated, no cookies. Every response carries `x-request-id`; 5xx bodies carry only `{error, requestId}`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
...
  "paths": {
    "/routes": {
      "get": {
        "tags": [
          "routes"
        ],
        "operationId": "listRoutes",
        "summary": "Routes sent, newest first",
        "parameters": [
          {
            "name": "ledger",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[-a-z0-9]{3,8}:[-_a-zA-Z0-9]{1,32}$",
              "example": "eip155:296"
            }
          },
          {
            "name": "sender",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            },
            "description": "Page size (max is `http.pagination.maxLimit`)."
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,512}$"
            },
            "description": "Opaque cursor: the `next` value of the previous page."
          }
        ],
        "responses": {
          "200": {
            "description": "A page of routes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/routes/{routeId}": {
      "get": {
        "tags": [
          "routes"
        ],
        "operationId": "getRoute",
        "summary": "Hop-by-hop status, tx hashes per ledger and the final outcome",
        "parameters": [
          {
            "name": "routeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0x[0-9a-fA-F]{32}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$"
            },
            "description": "16-byte hex route id or its UUID form (the UETR under the ISO 20022 filter)."
          }
        ],
        "responses": {
          "200": {
            "description": "Route status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteStatus"
                }
              }
            }
          },
          "404": {
            "description": "Route not seen (same body, `found: false`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteStatus"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/accounts/{caip10}/notices": {
      "get": {
        "tags": [
          "routes"
        ],
        "operationId": "getAccountNotices",
        "summary": "Quarantine notices, sender receipts and blacklist listings for an account",
        "parameters": [
          {
            "name": "caip10",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "CAIP-10 account id, URL-encoded."
          }
        ],
        "responses": {
          "200": {
            "description": "Notices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountNotices"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/registry": {
      "get": {
        "tags": [
          "registry"
        ],
        "operationId": "getRegistry",
        "summary": "Certifications, disabled routes, blacklist with case ids, committee and vault holdings",
        "parameters": [
          {
            "name": "ledger",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[-a-z0-9]{3,8}:[-_a-zA-Z0-9]{1,32}$",
              "example": "eip155:296"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registry views per ledger",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registry"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ledger",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/quote": {
      "post": {
        "tags": [
          "quotes"
        ],
        "operationId": "quote",
        "summary": "Planner quote over the live graph, with its inputs",
        "description": "Costs `http.rateLimit.quoteCost` rate-limit tokens. Bodies over `http.bodyLimitBytes` get 413. Unknown keys are rejected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote (`ok: true`) or no route (`ok: false` with reason)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Quote service busy or not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/graphs/{hash}": {
      "get": {
        "tags": [
          "quotes"
        ],
        "operationId": "getGraph",
        "summary": "Graph snapshot a quote was computed on (kept briefly)",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route graph snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Unknown or expired hash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/pending": {
      "get": {
        "tags": [
          "trigger"
        ],
        "operationId": "listPending",
        "summary": "Hops waiting for the public forward / flush trigger, with calldata anyone can send",
        "parameters": [
          {
            "name": "ledger",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[-a-z0-9]{3,8}:[-_a-zA-Z0-9]{1,32}$",
              "example": "eip155:296"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            },
            "description": "Page size (max is `http.pagination.maxLimit`)."
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,512}$"
            },
            "description": "Opaque cursor: the `next` value of the previous page."
          }
        ],
        "responses": {
          "200": {
            "description": "A page of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "retry-after": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait."
          }
        }
      },
      "Internal": {
        "description": "Internal error (no details; correlate with logs by requestId)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/InternalError"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Inputs": {
        "type": "object",
        "description": "Confirmed block per ledger the answer was built from.",
        "additionalProperties": {
          "$ref": "#/components/schemas/Cursor"
        }
      },
      "RoutePage": {
        "type": "object",
        "required": [
          "items",
          "next",
          "inputs"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "routeId",
                "ledger",
                "sender",
                "txHash",
                "blockNumber",
                "timestamp"
              ],
              "properties": {
                "routeId": {
                  "type": "string"
                },
                "ledger": {
                  "type": "string",
                  "pattern": "^[-a-z0-9]{3,8}:[-_a-zA-Z0-9]{1,32}$",
                  "example": "eip155:296"
                },
                "sender": {
                  "type": "string"
                },
                "destinationLedger": {
                  "type": "string"
                },
                "txHash": {
                  "type": "string",
                  "pattern": "^0x[0-9a-fA-F]{64}$"
                },
                "blockNumber": {
                  "type": "integer"
                },
                "timestamp": {
                  "type": "integer"
                }
              }
            }
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "inputs": {
            "$ref": "#/components/schemas/Inputs"
          }
        }
      },
...
    }
  }
}